Corrections: getters/setters, datatypes, controller

+++ backend/src/main/java/com/sdacademy/budgettracker/repository/ExpenseRepository.java
@@ -10,9 +10,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ExpenseRepository extends JpaRepository<Expense, Long> {
+public interface ExpenseRepository extends JpaRepository<Expense, Double> {
     @Query("SELECT SUM(e.amount) FROM Expense e WHERE e.user.id = :userId")
-    Optional<Double> findTotalExpensesByUserId(@Param("userId") Long userId);
+    Optional<Double> findTotalExpensesByUserId(@Param("userId") Double userId);
 
-    List<Expense> findAllByUserId(Long userId);
+    List<Expense> findAllByUserId(Double userId);
 }
