+++ backend/src/main/java/com/sdacademy/budgettracker/service/UserService.java
@@ -6,13 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
-
 @Service
 public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public User getCurrentUser(Long userId) {
+    public User getCurrentUser(Double userId) {
         return userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
     }
